fix: Move knb to sub-dir tools/knb
